Added sphere intersection and winning object intersection functions
Added a sphere intersection function that, given a ray, determines where that ray is coming from, and where on the object it's hitting. The winning object index function returns the index of the winning intersection aka the closest one to the camera. Fixed type issue where plain double intersection functions would not recognize intersections, needed to make them virtual doubles.